Set up routes in routers folder
