+++ src/app.js
@@ -1,55 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter, Link, NavLink, Route, Switch } from 'react-router-dom'
+import AppRouter from './routers/AppRouter';
 import 'normalize.css/normalize.css';
 import './styles/styles.scss';
 
-const ExpenseDashboardPage = () => (
-  <div>This is from my dashboard</div>
-)
-
-const AddExpensePage = () => (
-  <div>This is from my Add Expense</div>
-)
-
-const EditExpensePage = () => (
-  <div>This is from my Edit Expense</div>
-)
-
-const HelpPage = () => (
-  <div>This is from my HelpPage</div>
-)
-
-const NotFoundPage = () => (
-  <div>
-    404 - <Link to="/">Go Home</Link>
-  </div>
-)
-
-const Header = () => (
-  <header>
-    <h1>Budget App</h1>
-    <NavLink exact to="/" activeClassName="is-active" > Home</NavLink> |
-    <NavLink to="/create" activeClassName="is-active"> Create Expense</NavLink> |
-    <NavLink to="/edit" activeClassName="is-active"> Edit Expense</NavLink> |
-    <NavLink to="/help" activeClassName="is-active"> Help</NavLink>
-
-  </header>
-)
-
-const routes = (
-  <BrowserRouter>
-    <div>
-      <Header />
-      <Switch>
-        <Route exact path="/" component={ExpenseDashboardPage} />
-        <Route path="/create" component={AddExpensePage} />
-        <Route path="/edit" component={EditExpensePage} />
-        <Route path="/help" component={HelpPage} />
-        <Route component={NotFoundPage} />
-      </Switch>
-    </div>
-  </BrowserRouter>
-)
-
-ReactDOM.render(routes, document.getElementById('app'));
+ReactDOM.render(<AppRouter />, document.getElementById('app'));
